Rough game play works, you can open a game and input GPS data through the DDMS perspective on Eclipse.
Search a game by name works, and you can long click a found game and play it.

Like before, google key is not there so you have to enter your key manually. Also, to get the DDMS GPS functionality to work you have to change your Eclipse.ini file and add the following lines to it

-Duser.country=US
-Duser.language=en